fix(tasking): EPI-981: Build logic to clean-up repeating tasks on patient discharge
