+++ packages/shared/src/models/Discharge.js
@@ -6,7 +6,7 @@ import { buildEncounterLinkedSyncFilter } from './buildEncounterLinkedSyncFilter
 import { buildEncounterLinkedLookupFilter } from '../sync/buildEncounterLinkedLookupFilter';
 
 export class Discharge extends Model {
-  static init({ primaryKey, ...options }) {
+  static init({ primaryKey, ...options }, models) {
     const validate = {
       mustHaveEncounter() {
         if (!this.deletedAt && !this.encounterId) {
@@ -34,7 +34,16 @@ export class Discharge extends Model {
           allowNull: true,
         },
       },
-      { ...options, syncDirection: SYNC_DIRECTIONS.BIDIRECTIONAL, validate },
+      {
+        ...options,
+        syncDirection: SYNC_DIRECTIONS.BIDIRECTIONAL,
+        validate,
+        hooks: {
+          afterCreate: async (instance) => {
+            await models.Task.onEncounterDischarged(instance.encounterId);
+          },
+        },
+      },
     );
   }
